refactor: PieceView 와 PiecesView 를 사용하도록 변경
- 의도하지 않은 상태 변경을 방지하기 위함.
