+++ src/main/java/janggi/domain/piece/Piece.java
@@ -1,9 +1,10 @@
 package janggi.domain.piece;
 
 import janggi.domain.piece.movement.MovementStrategy;
+import janggi.domain.piece.pieces.PiecesView;
 import java.util.Objects;
 
-public class Piece {
+public class Piece implements PieceView {
 
     private final PieceType pieceType;
     private final Side side;
@@ -17,30 +18,34 @@ public class Piece {
         this.position = new Position(x, y);
     }
 
-    public final Side getSide() {
-        return side;
-    }
-
-    public final Position getPosition() {
-        return position;
-    }
-
-    public final boolean isSamePosition(Position position) {
-        return this.position.equals(position);
-    }
-
-    public void move(Pieces map, int x, int y) {
+    public void move(PiecesView map, int x, int y) {
         Position destination = new Position(x, y);
         validateMovable(map, destination);
         position = destination;
     }
 
-    private void validateMovable(Pieces map, Position destination) {
+    private void validateMovable(PiecesView map, Position destination) {
         if (!movementStrategy.isMoveable(map, position, side, destination)) {
             throw new IllegalArgumentException("해당 위치로 이동할 수 없습니다.");
         }
     }
 
+    @Override
+    public final Side getSide() {
+        return side;
+    }
+
+    @Override
+    public final Position getPosition() {
+        return position;
+    }
+
+    @Override
+    public final boolean isSamePosition(Position position) {
+        return this.position.equals(position);
+    }
+
+    @Override
     public PieceType getPieceType() {
         return pieceType;
     }
