+++ src/main/java/janggi/domain/piece/movement/MovementStrategy.java
@@ -1,10 +1,10 @@
 package janggi.domain.piece.movement;
 
-import janggi.domain.piece.Pieces;
+import janggi.domain.piece.pieces.PiecesView;
 import janggi.domain.piece.Position;
 import janggi.domain.piece.Side;
 
 public interface MovementStrategy {
 
-    boolean isMoveable(Pieces map, Position origin, Side side, Position destination);
+    boolean isMoveable(PiecesView map, Position origin, Side side, Position destination);
 }
