+++ src/main/java/janggi/domain/piece/movement/dynamic/CannonMovementStrategy.java
@@ -1,7 +1,8 @@
 package janggi.domain.piece.movement.dynamic;
 
 import janggi.domain.piece.PieceType;
-import janggi.domain.piece.Pieces;
+import janggi.domain.piece.pieces.Pieces;
+import janggi.domain.piece.pieces.PiecesView;
 import janggi.domain.piece.Position;
 import janggi.domain.piece.Side;
 import java.util.stream.IntStream;
@@ -17,26 +18,26 @@ public class CannonMovementStrategy extends DynamicMovementStrategy {
     }
 
     @Override
-    public boolean isLegalPath(Pieces existingPieces, Side side, Position origin, Position destination) {
-        Pieces piecesOnPath = getPiecesOnPath(existingPieces, origin, destination);
+    public boolean isLegalPath(PiecesView existingPieces, Side side, Position origin, Position destination) {
+        PiecesView piecesOnPath = getPiecesOnPath(existingPieces, origin, destination);
         if (piecesOnPath.size() != 1) {
             return false;
         }
         return !piecesOnPath.containsPieceType(PieceType.CANNON);
     }
 
-    private Pieces getPiecesOnPath(Pieces existingPieces, Position origin, Position destination) {
+    private PiecesView getPiecesOnPath(PiecesView existingPieces, Position origin, Position destination) {
         if (origin.hasSameX(destination)) {
             return getPiecesOnVerticalPath(existingPieces, origin, destination);
         }
         return getPiecesOnHorizontalPath(existingPieces, origin, destination);
     }
 
-    private Pieces getPiecesOnHorizontalPath(Pieces existingPieces, Position origin, Position destination) {
+    private PiecesView getPiecesOnHorizontalPath(PiecesView existingPieces, Position origin, Position destination) {
         int startX = Math.min(origin.getX(), destination.getX()) + 1;
         int endX = Math.max(origin.getX(), destination.getX());
 
-        return Pieces.from(
+        return Pieces.fromPieceViews(
             IntStream.range(startX, endX)
                 .mapToObj(x -> new Position(x, origin.getY()))
                 .flatMap(position -> existingPieces.findByPosition(position).stream())
@@ -44,11 +45,11 @@ public class CannonMovementStrategy extends DynamicMovementStrategy {
         );
     }
 
-    private Pieces getPiecesOnVerticalPath(Pieces existingPieces, Position origin, Position destination) {
+    private PiecesView getPiecesOnVerticalPath(PiecesView existingPieces, Position origin, Position destination) {
         int startY = Math.min(origin.getY(), destination.getY()) + 1;
         int endY = Math.max(origin.getY(), destination.getY());
 
-        return Pieces.from(
+        return Pieces.fromPieceViews(
             IntStream.range(startY, endY)
                 .mapToObj(y -> new Position(origin.getX(), y))
                 .flatMap(position -> existingPieces.findByPosition(position).stream())
