+++ src/test/java/janggi/domain/piece/movement/dynamic/CannonMovementStrategyTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import janggi.domain.ReplaceUnderBar;
 import janggi.domain.piece.Piece;
 import janggi.domain.piece.PieceType;
-import janggi.domain.piece.Pieces;
+import janggi.domain.piece.pieces.Pieces;
 import janggi.domain.piece.Position;
 import janggi.domain.piece.Side;
 import java.util.List;
